Move CondorLauncher methods from stops.go to condor.go. Rename handleEvents to routeEvents. Eliminate handlePing().
